Moving files in maven directories, updating pacakges. Using moody's pom config.
