+++ src/main/java/com/skedgo/converter/TimezoneMapper.java
@@ -1,6 +1,6 @@
 /** The provided code is written by Tim Cooper:   tim@edval.com.au
 This code is available under the MIT licence:  https://opensource.org/licenses/MIT  */
-
+package com.skedgo.converter;
 
 public class TimezoneMapper {
