+++ src/main/java/com/skedgo/generator/BoundingBox.java
@@ -1,4 +1,4 @@
-package com.skedgo;
+package com.skedgo.generator;
 
 import org.jetbrains.annotations.NotNull;
