+++ src/main/java/com/skedgo/generator/BoundingBoxOrPolygon.java
@@ -1,4 +1,4 @@
-package com.skedgo;
+package com.skedgo.generator;
 
 
 import java.util.List;
