+++ src/main/java/com/skedgo/generator/Parser.java
@@ -1,4 +1,4 @@
-package com.skedgo;
+package com.skedgo.generator;
