+++ src/main/java/com/skedgo/generator/parsing/JSONObject.java
@@ -1,4 +1,4 @@
-package com.skedgo.Parsing;
+package com.skedgo.generator.parsing;
 
 /*
 Copyright (c) 2002 JSON.org
