+++ src/main/java/com/skedgo/generator/parsing/JSONTokener.java
@@ -1,4 +1,4 @@
-package com.skedgo.Parsing;
+package com.skedgo.generator.parsing;
 
 
 import java.io.BufferedReader;
